+++ sync-for-payables/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v1.1.0] sync-for-payables
 ### Releases
 - [PyPI v1.1.0] https://pypi.org/project/codat-sync-for-payables/1.1.0 - sync-for-payables
+
+## 2023-09-22 15:32:22
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.90.0 (2.125.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.2.0] sync-for-payables
+### Releases
+- [PyPI v1.2.0] https://pypi.org/project/codat-sync-for-payables/1.2.0 - sync-for-payables